% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/added-variable-plot.R
\name{addvar_plot}
\alias{addvar_plot}
\title{Added Variable Plot}
\usage{
addvar_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Added variable plot provides information about the marginal importance of a 
predictor variable, given the other predictor variables already in 
the model. It shows the marginal importance of the variable in reducing the
residual variability.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
addvar_plot(model)
}
\references{
Chatterjee, Samprit and Hadi, Ali. Regression Analysis by Example. 5th ed. N.p.: John Wiley & Sons, 2012. Print.
Kutner, MH, Nachtscheim CJ, Neter J and Li W., 2004, Applied Linear Statistical Models (5th edition). 
Chicago, IL., McGraw Hill/Irwin.
}

