% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-criteria.R
\name{aic}
\alias{aic}
\title{AIC}
\usage{
aic(model, method = c("R", "STATA", "SAS"))
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{method}{a character vector; specify the method to compute AIC. Valid
options include R, STATA and SAS}
}
\value{
Akaike Information Criteria
}
\description{
Akaike Information Criteria
}
\details{
Some statistical tests, for example the analysis of variance, assume
that variances are equal across groups or samples. The Bartlett test can be
used to verify that assumption. Bartlett's test is sensitive to departures
from normality. That is, if your samples come from non-normal distributions,
then Bartlett's test may simply be testing for non-normality. The Levene test
is an alternative to the Bartlett test that is less sensitive to departures
from normality.
}

