% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-subsets.R
\name{all_subset}
\alias{all_subset}
\title{All Subsets Regression}
\usage{
all_subset(model, ...)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{...}{other arguments}
}
\value{
\code{all_subset} returns an object of class \code{"all_subset"}.
An object of class \code{"all_subset"} is a data frame containing the
following components:

\item{n}{model number}
\item{predictors}{predictors in the model}
\item{rsquare}{rsquare of the model}
\item{adjr}{adjusted rsquare of the model}
\item{predrsq}{predicted rsquare of the model}
\item{cp}{Mallow's Cp}
\item{aic}{Akaike Information Criteria}
\item{sbic}{Sawa Bayesian Information Criteria}
\item{sbc}{Schwarz Bayes Information Criteria}
\item{gmsep}{estimated MSE of prediction, assuming multivariate normality}
\item{jp}{final prediction error}
\item{pc}{Amemiya Prediction Criteria}
\item{sp}{Hocking's Sp}
}
\description{
Performs all subset regression for a linear regression model.
}
\examples{
model <- lm(mpg ~ disp + hp, data = mtcars)
all_subset(model)

}

