% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-data-auto.R
\docType{data}
\name{auto}
\alias{auto}
\title{Test Data Set}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 11 columns.}
\usage{
auto
}
\description{
Test Data Set
}
\keyword{datasets}
