% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best-subsets.R
\name{best_subset}
\alias{best_subset}
\title{Best Subsets Regression}
\usage{
best_subset(model, ...)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{...}{other inputs}
}
\value{
\code{best_subset} returns an object of class \code{"best_subset"}.
An object of class \code{"best_subset"} is a data frame containing the
following components:

\item{n}{model number}
\item{predictors}{predictors in the model}
\item{rsquare}{rsquare of the model}
\item{adjr}{adjusted rsquare of the model}
\item{predrsq}{predicted rsquare of the model}
\item{cp}{Mallow's Cp}
\item{aic}{Akaike Information Criteria}
\item{sbic}{Sawa Bayesian Information Criteria}
\item{sbc}{Schwarz Bayes Information Criteria}
\item{gmsep}{estimated MSE of prediction, assuming multivariate normality}
\item{jp}{final prediction error}
\item{pc}{Amemiya Prediction Criteria}
\item{sp}{Hocking's Sp}
}
\description{
Select the subset of predictors that do the best at meeting some
well-defined objective criterion, such as having the largest R2 value or the
smallest MSE.
}
\details{
Some statistical tests, for example the analysis of variance, assume
that variances are equal across groups or samples. The Bartlett test can be
used to verify that assumption. Bartlett's test is sensitive to departures
from normality. That is, if your samples come from non-normal distributions,
then Bartlett's test may simply be testing for non-normality. The Levene test
is an alternative to the Bartlett test that is less sensitive to departures
from normality.
}

