% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box-resid.R
\name{box_resid}
\alias{box_resid}
\title{Residual Box Plot}
\usage{
box_resid(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Box plot of residuals from a linear regression model
}

