% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cement.R
\docType{data}
\name{cement}
\alias{cement}
\title{Cement Data Set}
\format{A data frame with 13 rows and 6 variables:
\describe{
  \item{X}{id}
  \item{y}{y}
  \item{x1}{x1}
  \item{x2}{x2}
  \item{x3}{x3}
  \item{x4}{x4}
}}
\source{
\url{http://www.ats.ucla.edu/stat/spss/whatstat/whatstat.htm}
}
\usage{
cement
}
\description{
A dataset containing demographic information about
cement manufacturing.
}
\keyword{datasets}

