% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coll-diag.R
\name{coll_diag}
\alias{coll_diag}
\alias{eigen_cindex}
\alias{vif_tol}
\title{Collinearity Diagnostics}
\usage{
coll_diag(model)

vif_tol(model)

eigen_cindex(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{coll_diag} returns an object of class \code{"coll_diag"}.
An object of class \code{"coll_diag"} is a list containing the
following components:

\item{vif_t}{tolerance and variance inflation factors}
\item{eig_cindex}{eigen values and condition index}
}
\description{
Test if k samples have equal variances
}

