% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted-line-properties.R
\name{conf_band}
\alias{conf_band}
\title{Confidence Band Regression Line}
\usage{
conf_band(model, conf.level = 0.95)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{conf.level}{confidence level}
}
\value{
\code{conf_band} returns  a data frame
}
\description{
confidence band for regression line
}

