% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted-line-properties.R
\name{conf_line}
\alias{conf_line}
\title{Confidence Interval Regression Line}
\usage{
conf_line(model, conf.level = 0.95)
}
\arguments{
\item{model}{a simple linear regression model}

\item{conf.level}{confidence level for the interval}
}
\value{
\code{reg_line} returns  a list containing the
following components:

\item{conf_band}{confidence band for the regression line}
}
\description{
Confidence interval for a simple linear regression line
}

