% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks-d-barplot.R
\name{cooksd_bplot}
\alias{cooksd_bplot}
\title{Cooks' d Bar Plot}
\usage{
cooksd_bplot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{cooksd_bplot} returns a list containing the
following components:

\item{cooks_d}{f cook's d statistic}
\item{threshold}{threshold for outliers}
\item{normal}{normal residual points}
\item{outlier}{residual points that are outliers}
}
\description{
Cooks' d Bar Plot
}

