% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks-d-chart.R
\name{cooksd_chart}
\alias{cooksd_chart}
\title{Cooks' D Chart}
\usage{
cooksd_chart(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Chart of Cook's distance to detect observations that strongly influence fitted values of the model.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
cooksd_chart(model)
}

