% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlations}
\alias{correlations}
\title{Part and Partial Correlations}
\usage{
correlations(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{correlations} returns an object of class \code{"correlations"}.
An object of class \code{"correlations"} is a data frame containing the
following components:

\item{Zero-order}{zero order correlations}
\item{Partial}{partial correlations}
\item{Part}{part correlations}
}
\description{
Zero-order, part and partial correlations
}

