% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffits-plot.R
\name{dffits_plot}
\alias{dffits_plot}
\title{DFFITS Plot}
\usage{
dffits_plot(model)
}
\description{
Plot for detecting influential observations using DFFITS.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
dffits_panel(model)
}

