% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-panel.R
\name{diag_panel}
\alias{diag_panel}
\title{Diagnostics Panel}
\usage{
diag_panel(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Panel of plots for regression diagnostics
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
diag_panel(model)
}

