% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsresid-vs-pred-plot.R
\name{dsrvsp_plot}
\alias{dsrvsp_plot}
\title{Deleted Studentized Residual vs Fitted Values Plot}
\usage{
dsrvsp_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Deleted Studentized Residual vs Fitted Values Plot
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
dsrvsp_plot(model)
}

