% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\docType{data}
\name{fitness}
\alias{fitness}
\title{Fitness Data Set}
\format{A data frame with 31 rows and 7 variables:
\describe{
  \item{age}{id of the student}
  \item{weight}{gender of the student}
  \item{oxygen}{ethnic background of the student}
  \item{runtime}{socio-economic status of the student}
  \item{restpulse}{school type}
  \item{runpulse}{program type}
  \item{maxpulse}{scores from test of reading}
}}
\source{
\url{http://www.ats.ucla.edu/stat/spss/whatstat/whatstat.htm}
}
\usage{
fitness
}
\description{
A dataset containing fitness metrics
}
\keyword{datasets}

