% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid-fit-spread-plot.R
\name{fm_plot}
\alias{fm_plot}
\title{Fit Mean Plot}
\usage{
fm_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{fm_plot} returns a list containing the
following components:

\item{fitted_values}{f cook's d statistic}
\item{avg_fitted}{threshold for outliers}
\item{dev_fitted}{f cook's d statistic}
\item{ecdf_avg_fitted}{threshold for outliers}
}
\description{
Fit Mean Plot
}

