% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-criteria.R
\name{gmsep}
\alias{gmsep}
\title{GMSEP}
\usage{
gmsep(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
GMSEP
}
\description{
Estimated error of prediction, assuming multivariate normality
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
gmsep(model)
}

