% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence-measures.R
\name{hadi}
\alias{hadi}
\title{Hadi's Measure}
\usage{
hadi(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
hadi's measure
}
\description{
Measure of influence based on the fact that influential observations in either
the response variable or in the predictors or both.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
hadi_plot(model)
}

