% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual-hist.R
\name{hist_resid}
\alias{hist_resid}
\title{Residual Histogram}
\usage{
hist_resid(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Histogram of residuals for detecting violation of normality assumption.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
hist_resid(model)
}

