% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-criteria.R
\name{jp}
\alias{jp}
\title{JP}
\usage{
jp(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
JP
}
\description{
Final prediction error
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
jp(model)
}

