% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-criteria.R
\name{mallow_cp}
\alias{mallow_cp}
\title{Mallow's Cp}
\usage{
mallow_cp(model, fullmodel)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{fullmodel}{an object of class \code{lm}}
}
\value{
Mallow's Cp
}
\description{
Mallow's Cp
}
\examples{
full_model <- lm(mpg ~ ., data = mtcars)
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
mallow_cp(model, full_model)
}

