% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm-test.R
\name{norm_test}
\alias{norm_test}
\title{Normality Test}
\usage{
norm_test(y, ...)
}
\arguments{
\item{y}{a numeric vector}

\item{...}{other arguments}
}
\value{
\code{norm_test} returns an object of class \code{"norm_test"}.
An object of class \code{"norm_test"} is a list containing the
following components:

\item{kolmogorv}{kolmogorv smirnov statistic}
\item{shapiro}{shapiro wilk statistic}
\item{cramer}{cramer von mises statistic}
\item{anderson}{anderson darling statistic}
}
\description{
Normality Test
}
\details{
Some statistical tests, for example the analysis of variance, assume
that variances are equal across groups or samples. The Bartlett test can be
used to verify that assumption. Bartlett's test is sensitive to departures
from normality. That is, if your samples come from non-normal distributions,
then Bartlett's test may simply be testing for non-normality. The Levene test
is an alternative to the Bartlett test that is less sensitive to departures
from normality.
}

