% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-all-possible-regression.R
\name{ols_all_subset_betas}
\alias{ols_all_subset_betas}
\title{All Possible Regression Variable Coefficients}
\usage{
ols_all_subset_betas(object, ...)
}
\arguments{
\item{object}{an object of class \code{lm}}

\item{...}{other arguments}
}
\value{
\code{ols_all_subset_betas} returns a tibble containing 

\item{model_index}{model number}
\item{predictor}{predictor}
\item{beta_coef}{coefficient for the predictor}
}
\description{
Returns the coefficients for each variable from each model
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_all_subset_betas(model)

}
