% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-anova.R
\name{ols_anova}
\alias{ols_anova}
\title{ANOVA}
\usage{
ols_anova(model, ...)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{...}{Other arguments.}
}
\value{
\code{ols_anova} returns an object of class \code{"ols_anova"}.
An object of class \code{"ols_anova"} is a list containing the following
components:

\item{error_df}{residual degrees of freedom}
\item{model_df}{regression degrees of freedom}
\item{total_df}{total degrees of freedom}
\item{ess}{error sum of squares}
\item{rss}{regression sum of squares}
\item{tss}{total sum of squares}
\item{rms}{regression mean square}
\item{ems}{error mean square}
\item{f}{f statistis}
\item{p}{p-value for \code{f}}
}
\description{
ANOVA for regression model.
}
\examples{
# model
model <- lm(mpg ~ disp + hp + wt, data = mtcars)

# anova
ols_anova(model)

}
