% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-information-criteria.R
\name{ols_apc}
\alias{ols_apc}
\title{Amemiya's Prediction Criteria}
\usage{
ols_apc(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
Amemiya's prediction error
}
\description{
Amemiya's prediction error
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_apc(model)
}
\references{
Amemiya, T. (1976). Selection of Regressors. Technical Report 225, Stanford University, Stanford, CA.

Judge, G. G., Griffiths, W. E., Hill, R. C., and Lee, T.-C. (1980). The Theory and Practice of Econometrics.
New York: John Wiley & Sons.
}

