% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-performance.R
\name{ols_compare_model_performance}
\alias{ols_compare_model_performance}
\title{Compare model performance}
\usage{
ols_compare_model_performance(...)
}
\arguments{
\item{...}{Multiple objects of class \code{lm}.}
}
\description{
Compare performance of regression models using multiple metrics.
}
\examples{
model_1 <- lm(mpg ~ disp + hp, data = mtcars)
model_2 <- lm(mpg ~ disp + hp + wt, data = mtcars)
model_3 <- lm(mpg ~ disp + hp + wt + qsec + + drat, data = mtcars)
ols_compare_model_performance(model_1, model_2, model_3)

}
