% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-cooks-d-barplot.R
\name{ols_cooksd_barplot}
\alias{ols_cooksd_barplot}
\title{Cooks' D Bar Plot}
\usage{
ols_cooksd_barplot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Bar Plot of Cook's distance to detect observations that strongly influence fitted values of the model.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_cooksd_barplot(model)
}

