% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-normality-tests.R
\name{ols_corr_test}
\alias{ols_corr_test}
\title{Correlation Test For Normality}
\usage{
ols_corr_test(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
correlation between fitted regression model residuals and expected
values of residuals
}
\description{
Correlation between observed residuals and expected residuals under normality.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_corr_test(model)
}

