% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_count_obs}
\alias{ols_count_obs}
\title{Observations}
\usage{
ols_count_obs(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Number of observations.
}
\description{
Returns the number of observation from a model.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_count_obs(model)

}
