% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_count_parameters}
\alias{ols_count_parameters}
\title{Parameters count}
\usage{
ols_count_parameters(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Number of parameters.
}
\description{
Returns the number of parameters from a model.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_count_parameters(model)

}
