% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-dfbetas-panel.R
\name{ols_dfbetas_panel}
\alias{ols_dfbetas_panel}
\title{DFBETAs Panel}
\usage{
ols_dfbetas_panel(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Panel of plots to detect influential observations using DFBETAs.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_dfbetas_panel(model)
}

