% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-dffits-plot.R
\name{ols_dffits_plot}
\alias{ols_dffits_plot}
\title{DFFITS Plot}
\usage{
ols_dffits_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Plot for detecting influential observations using DFFITS.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_dffits_plot(model)
}

