% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-diagnostics-panel.R
\name{ols_diagnostic_panel}
\alias{ols_diagnostic_panel}
\title{Diagnostics Panel}
\usage{
ols_diagnostic_panel(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Panel of plots for regression diagnostics
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_diagnostic_panel(model)
}

