% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-residual-fit-spread-plot.R
\name{ols_fm_plot}
\alias{ols_fm_plot}
\title{Fit Mean Plot}
\usage{
ols_fm_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Fit Mean Plot
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_fm_plot(model)
}

