% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_call}
\alias{ols_get_call}
\title{Model function call}
\usage{
ols_get_call(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Model's function call.
}
\description{
Returns the model's function call.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_call(model)

}
