% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_data}
\alias{ols_get_data}
\title{Model data}
\usage{
ols_get_data(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Returns data used in model.
}
\examples{
model <- lm(mpg ~ wt + cyl * hp * disp + gear * drat, data = mtcars)
ols_get_data(model)

}
