% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_deviance}
\alias{ols_get_deviance}
\title{Deviance}
\usage{
ols_get_deviance(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Model deviance.
}
\description{
Returns the model deviance.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_deviance(model)

}
