% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_df}
\alias{ols_get_df}
\title{Degrees of freedom}
\usage{
ols_get_df(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Residual degrees of freedom.
}
\description{
Returns residual degrees of freedom.
}
\examples{
model <- lm(mpg ~ wt + cyl * hp * disp + gear * drat, data = mtcars)
ols_get_df(model)

}
