% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_formula}
\alias{ols_get_formula}
\title{Model formula}
\usage{
ols_get_formula(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
An object of class \code{formula}.
}
\description{
Returns the model formula.
}
\examples{
# model
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_get_formula(model)

}
\concept{model info}
