% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_interaction_terms}
\alias{ols_get_interaction_terms}
\title{Interaction terms}
\usage{
ols_get_interaction_terms(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
\code{Character} vector or \code{NULL}.
}
\description{
Returns interaction terms present in the model.
}
\examples{
model <- lm(mpg ~ wt + cyl * hp * disp + gear * drat, data = mtcars)
ols_get_interaction_terms(model)

}
