% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_intercept}
\alias{ols_get_intercept}
\title{Model intercept}
\usage{
ols_get_intercept(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
The value of the intercept.
}
\description{
Returns the value of the model intercept.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_intercept(model)

model <- lm(mpg ~ 0 + wt + cyl + hp + disp, data = mtcars)
ols_get_intercept(model)

}
