% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_model_matrix}
\alias{ols_get_model_matrix}
\title{Model matrix}
\usage{
ols_get_model_matrix(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
A \code{data.frame}.
}
\description{
Returns the design matrix.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_model_matrix(model)

}
