% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_parameters}
\alias{ols_get_parameters}
\title{Model parameters}
\usage{
ols_get_parameters(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
A \code{data.frame} with model coefficients.
}
\description{
Returns the coefficients.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_parameters(model)

}
