% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_predicted}
\alias{ols_get_predicted}
\title{Fitted values}
\usage{
ols_get_predicted(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Fitted values.
}
\description{
Returns the fitted values.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_predicted(model)

}
