% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_predictors}
\alias{ols_get_predictors}
\title{Predictor variable data}
\usage{
ols_get_predictors(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
A \code{data.frame} with model predictors.
}
\description{
Returns the data of all model predictors.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_predictors(model)

}
