% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_residuals}
\alias{ols_get_residuals}
\title{Model residuals}
\usage{
ols_get_residuals(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Residuals from the regression model.
}
\description{
Returns the residuals from the model.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_residuals(model)

}
