% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_response}
\alias{ols_get_response}
\title{Response variable data}
\usage{
ols_get_response(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Values of the response variable.
}
\description{
Returns the values of the response variable of a model.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_response(model)

}
