% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_sigma}
\alias{ols_get_sigma}
\title{Sigma}
\usage{
ols_get_sigma(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Residual standard deviation.
}
\description{
Returns the residual standard deviation.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_sigma(model)

}
