% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_variables}
\alias{ols_get_variables}
\title{Model variables}
\usage{
ols_get_variables(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
List with response and predictors.
}
\description{
Return model variables.
}
\examples{
model <- lm(mpg ~ wt + cyl * hp * disp + gear * drat, data = mtcars)
ols_get_variables(model)

}
