% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_get_vcov}
\alias{ols_get_vcov}
\title{Variance covariance matrix}
\usage{
ols_get_vcov(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Variance covariance matrix.
}
\description{
Returns the variance covariance matrix from the model.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_get_vcov(model)

}
