% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-information-criteria.R
\name{ols_gmsep}
\alias{ols_gmsep}
\title{GMSEP}
\usage{
ols_gmsep(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
GMSEP
}
\description{
Estimated error of prediction, assuming multivariate normality
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_gmsep(model)
}

