% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_has_intercept}
\alias{ols_has_intercept}
\title{Model intercept check}
\usage{
ols_has_intercept(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
\code{TRUE} if \code{model} has an intercept, \code{FALSE} otherwise.
}
\description{
Checks if model has an intercept.
}
\examples{
model <- lm(mpg ~ wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_has_intercept(model)

model <- lm(mpg ~ 0 + wt + cyl + hp + disp + gear + drat, data = mtcars)
ols_has_intercept(model)

}
