% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-information-criteria.R
\name{ols_jp}
\alias{ols_jp}
\title{JP}
\usage{
ols_jp(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
JP
}
\description{
Final prediction error
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_jp(model)
}

