% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-launch-app.R
\name{ols_launch_app}
\alias{ols_launch_app}
\title{Launch shiny app}
\usage{
ols_launch_app()
}
\description{
Launches shiny app for interactive model building.
}
\examples{
\dontrun{
ols_launch_app()
}
}
