% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-info.R
\name{ols_model_info}
\alias{ols_model_info}
\title{Model Info}
\usage{
ols_model_info(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
A list with information about model objects.
}
\description{
Comprehensive information from model objects.
}
\examples{
model <- lm(mpg ~ wt + cyl * hp * disp + gear * drat, data = mtcars)
ols_model_info(model)

}
