% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-model-performance.R
\name{ols_model_performance}
\alias{ols_model_performance}
\title{Model Performance}
\usage{
ols_model_performance(model, ...)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{...}{Other arguments.}
}
\value{
\code{ols_model_performance} returns an object of class
\code{"ols_model_performance"}. An object of class \code{"ols_regress"} is a
list containing the following components:

\item{r}{square root of rsquare, correlation between observed and predicted values of dependent variable}
\item{rsq}{coefficient of determination or r-square}
\item{adjr}{adjusted rsquare}
\item{prsq}{predicted rsquare}
\item{aic}{akaike information criteria}
\item{sbc}{bayesian information criteria}
\item{sbic}{sawa bayesian information criteria}
\item{rmse}{root mean squared error}
}
\description{
Evaluate performance of regression models.
}
\examples{
# model
model <- lm(mpg ~ disp + hp + wt, data = mtcars)

# model performance
ols_model_performance(model)

}
