% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-observed-vs-predicted-plot.R
\name{ols_ovsp_plot}
\alias{ols_ovsp_plot}
\title{Observed vs Predicted Plot}
\usage{
ols_ovsp_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Plot of observed vs fitted values to assess the fit of the model.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_ovsp_plot(model)
}

