% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-information-criteria.R
\name{ols_pc}
\alias{ols_pc}
\title{PC}
\usage{
ols_pc(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
PC
}
\description{
Amemiya's prediction error
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_pc(model)
}

