% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-fitted-line-properties.R
\name{ols_plot_reg_line}
\alias{ols_plot_reg_line}
\alias{ols_reg_line}
\title{Simple linear regression line}
\usage{
ols_plot_reg_line(response, predictor, print_plot = TRUE)
}
\arguments{
\item{response}{Response variable.}

\item{predictor}{Predictor variable.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\description{
Plot to demonstrate that the regression line always passes  through mean of
the response and predictor variables.
}
\section{Deprecated Function}{

\code{ols_reg_line()} has been deprecated. Instead use \code{ols_plot_reg_line()}.
}

\examples{
ols_plot_reg_line(mtcars$mpg, mtcars$disp)

}
