% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-residual-histogram.R
\name{ols_plot_resid_hist}
\alias{ols_plot_resid_hist}
\alias{ols_rsd_hist}
\title{Residual histogram}
\usage{
ols_plot_resid_hist(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Histogram of residuals for detecting violation of normality assumption.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_plot_resid_hist(model)

}
\seealso{
Other residual diagnostics: \code{\link{ols_plot_resid_box}},
  \code{\link{ols_rsd_qqplot}},
  \code{\link{ols_rvsp_plot}},
  \code{\link{ols_test_correlation}},
  \code{\link{ols_test_normality}}
}
