% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-rstud-vs-lev-plot.R
\name{ols_plot_resid_lev}
\alias{ols_plot_resid_lev}
\alias{ols_rsdlev_plot}
\title{Studentized residuals vs leverage plot}
\usage{
ols_plot_resid_lev(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Graph for detecting outliers and/or observations with high leverage.
}
\section{Deprecated Function}{

\code{ols_rsdlev_plot()} has been deprecated. Instead use \code{ols_plot_resid_lev()}.
}

\examples{
model <- lm(read ~ write + math + science, data = hsb)
ols_plot_resid_lev(model)

}
\seealso{
[ols_plot_resid_stud_fit()], [ols_plot_resid_lev()]
}
