% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-response-profile.R
\name{ols_plot_response}
\alias{ols_plot_response}
\title{Response variable profile}
\usage{
ols_plot_response(model, print_plot = TRUE)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\description{
Panel of plots to explore and visualize the response variable.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_plot_response(model)

}
