% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-plots-data.R
\name{ols_prep_avplot_data}
\alias{ols_prep_avplot_data}
\title{Added variable plot data}
\usage{
ols_prep_avplot_data(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Data for generating the added variable plots.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_prep_avplot_data(model)

}
