% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-plots-data.R
\name{ols_prep_dsrvf_data}
\alias{ols_prep_dsrvf_data}
\title{Deleted studentized residual plot data}
\usage{
ols_prep_dsrvf_data(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Generates data for deleted studentized residual vs fitted plot.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_prep_dsrvf_data(model)

}
