% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-plots-data.R
\name{ols_prep_outlier_obs}
\alias{ols_prep_outlier_obs}
\title{Cooks' D outlier observations}
\usage{
ols_prep_outlier_obs(k)
}
\arguments{
\item{k}{Cooks' d bar plot data.}
}
\description{
Identify outliers in cook's d plot.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
k <- ols_prep_cdplot_data(model)
ols_prep_outlier_obs(k)

}
