% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-plots-data.R
\name{ols_prep_regress_x}
\alias{ols_prep_regress_x}
\title{Regress predictor on other predictors}
\usage{
ols_prep_regress_x(data, i)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{i}{A numeric vector (indicates the predictor in the model).}
}
\description{
Regress a predictor in the model on all the other predictors.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
data <- ols_prep_avplot_data(model)
ols_prep_regress_x(data, 1)

}
