% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-plots-data.R
\name{ols_prep_regress_y}
\alias{ols_prep_regress_y}
\title{Regress y on other predictors}
\usage{
ols_prep_regress_y(data, i)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{i}{A numeric vector (indicates the predictor in the model).}
}
\description{
Regress y on all the predictors except the ith predictor.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
data <- ols_prep_avplot_data(model)
ols_prep_regress_y(data, 1)

}
