% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-plots-data.R
\name{ols_prep_rfsplot_fmdata}
\alias{ols_prep_rfsplot_fmdata}
\alias{ols_prep_rfsplot_rsdata}
\title{Residual fit spread plot data}
\usage{
ols_prep_rfsplot_fmdata(model)

ols_prep_rfsplot_rsdata(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Data for generating residual fit spread plot.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_prep_rfsplot_fmdata(model)
ols_prep_rfsplot_rsdata(model)

}
