% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-plots-data.R
\name{ols_prep_rstudlev_data}
\alias{ols_prep_rstudlev_data}
\title{Studentized residual vs leverage plot data}
\usage{
ols_prep_rstudlev_data(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Generates data for studentized resiudual vs leverage plot.
}
\examples{
model <- lm(read ~ write + math + science, data = hsb)
ols_prep_rstudlev_data(model)


}
