% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-plots-data.R
\name{ols_prep_rvsrplot_data}
\alias{ols_prep_rvsrplot_data}
\title{Residual vs regressor plot data}
\usage{
ols_prep_rvsrplot_data(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Data for generating residual vs regressor plot.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_prep_rvsrplot_data(model)

}
