% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-plots-data.R
\name{ols_prep_srplot_data}
\alias{ols_prep_srplot_data}
\title{Studentized residual plot data}
\usage{
ols_prep_srplot_data(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Generates data for studentized residual plot.
}
\examples{
model <- lm(read ~ write + math + science, data = hsb)
ols_prep_srplot_data(model)

}
