% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-fitted-line-properties.R
\name{ols_reg_line}
\alias{ols_reg_line}
\title{Simple Linear Regression Line}
\usage{
ols_reg_line(response, predictor)
}
\arguments{
\item{response}{response variable}

\item{predictor}{explanatory variable}
}
\description{
Regression line always passes through xbar and ybar
}
\examples{
ols_reg_line(mtcars$mpg, mtcars$disp)
}
