% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-residual-histogram.R
\name{ols_resid_hist}
\alias{ols_resid_hist}
\title{Residual Histogram}
\usage{
ols_resid_hist(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Histogram of residuals for detecting violation of normality assumption.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_resid_hist(model)
}

