% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-response-profile.R
\name{ols_resp_viz}
\alias{ols_resp_viz}
\title{Response variable profile}
\usage{
ols_resp_viz(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Panel of plots to explore and visualize the response variable.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_resp_viz(model)

}
