% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-residual-fit-spread-plot.R
\name{ols_rfs_plot}
\alias{ols_fm_plot}
\alias{ols_rfs_plot}
\alias{ols_rsd_plot}
\title{Residual Fit Spread Plot}
\usage{
ols_rfs_plot(model)

ols_fm_plot(model)

ols_rsd_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Plot to detect non-linearity, influential observations and outliers.
}
\examples{
# model
model <- lm(mpg ~ disp + hp + wt, data = mtcars)

# residual fit spread plot
ols_rfs_plot(model)

# fit mean plot
ols_fm_plot(model)

# residual spread plot
ols_rsd_plot(model)

}
\references{
Cleveland, W. S. (1993). Visualizing Data. Summit, NJ: Hobart Press.
}

