% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-residual-fit-spread-plot.R
\name{ols_rsd_plot}
\alias{ols_rsd_plot}
\title{Residual Spread Plot}
\usage{
ols_rsd_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Residual Spread Plot
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_rsd_plot(model)
}

