% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-residual-vs-predicted-plot.R
\name{ols_rvsp_plot}
\alias{ols_rvsp_plot}
\title{Residual vs Fitted Plot}
\usage{
ols_rvsp_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Plot to detect non-linearity, unequal error variances, and outliers.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_rvsp_plot(model)
}

