% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-information-criteria.R
\name{ols_sp}
\alias{ols_sp}
\title{SP}
\usage{
ols_sp(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
SP
}
\description{
Average prediction mean squared error
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_sp(model)
}

