% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-studentized-residual-chart.R
\name{ols_srsd_chart}
\alias{ols_srsd_chart}
\title{Standardized Residual Chart}
\usage{
ols_srsd_chart(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Chart for identifying outliers
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_srsd_chart(model)
}

