% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-studentized-residual-plot.R
\name{ols_srsd_plot}
\alias{ols_srsd_plot}
\title{Studentized Residual Plot}
\usage{
ols_srsd_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Graph for identifying outliers
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_srsd_plot(model)
}

