% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-stepaic-backward-regression.R
\name{ols_step_backward_aic}
\alias{ols_step_backward_aic}
\alias{ols_step_backward_aic.default}
\alias{plot.ols_step_backward_aic}
\alias{ols_step_backward_sbc}
\alias{ols_step_backward_sbc.default}
\alias{plot.ols_step_backward_sbc}
\alias{ols_step_backward_sbic}
\alias{ols_step_backward_sbic.default}
\alias{plot.ols_step_backward_sbic}
\alias{ols_step_backward_r2}
\alias{ols_step_backward_r2.default}
\alias{plot.ols_step_backward_r2}
\alias{ols_step_backward_adj_r2}
\alias{ols_step_backward_adj_r2.default}
\alias{plot.ols_step_backward_adj_r2}
\title{Stepwise IC backward regression}
\usage{
ols_step_backward_aic(model, ...)

\method{ols_step_backward_aic}{default}(
  model,
  include = NULL,
  exclude = NULL,
  progress = FALSE,
  details = FALSE,
  ...
)

\method{plot}{ols_step_backward_aic}(x, print_plot = TRUE, details = TRUE, digits = 3, ...)

ols_step_backward_sbc(model, ...)

\method{ols_step_backward_sbc}{default}(
  model,
  include = NULL,
  exclude = NULL,
  progress = FALSE,
  details = FALSE,
  ...
)

\method{plot}{ols_step_backward_sbc}(x, print_plot = TRUE, details = TRUE, digits = 3, ...)

ols_step_backward_sbic(model, ...)

\method{ols_step_backward_sbic}{default}(
  model,
  include = NULL,
  exclude = NULL,
  progress = FALSE,
  details = FALSE,
  ...
)

\method{plot}{ols_step_backward_sbic}(x, print_plot = TRUE, details = TRUE, digits = 3, ...)

ols_step_backward_r2(model, ...)

\method{ols_step_backward_r2}{default}(
  model,
  include = NULL,
  exclude = NULL,
  progress = FALSE,
  details = FALSE,
  ...
)

\method{plot}{ols_step_backward_r2}(x, print_plot = TRUE, details = TRUE, digits = 3, ...)

ols_step_backward_adj_r2(model, ...)

\method{ols_step_backward_adj_r2}{default}(
  model,
  include = NULL,
  exclude = NULL,
  progress = FALSE,
  details = FALSE,
  ...
)

\method{plot}{ols_step_backward_adj_r2}(x, print_plot = TRUE, details = TRUE, digits = 3, ...)
}
\arguments{
\item{model}{An object of class \code{lm}; the model should include all
candidate predictor variables.}

\item{...}{Other arguments.}

\item{include}{Character or numeric vector; variables to be included in selection process.}

\item{exclude}{Character or numeric vector; variables to be excluded from selection process.}

\item{progress}{Logical; if \code{TRUE}, will display variable selection progress.}

\item{details}{Logical; if \code{TRUE}, will print the regression result at
each step.}

\item{x}{An object of class \code{ols_step_backward_*}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\value{
\code{ols_step_backward_*} returns an object of class \code{"ols_step_backward_*"}.
An object of class \code{"ols_step_backward_*"} is a list containing the
following components:

\item{model}{final model; an object of class \code{lm}}
\item{metrics}{selection metrics}
\item{others}{list; info used for plotting and printing}
}
\description{
Build regression model from a set of candidate predictor variables by
removing predictors based on akaike information criterion, schwarz bayesian criterion,
sawa bayesian criterion, r-squared or adjusted r-squared, in a stepwise
manner until there is no variable left to remove any more.
}
\examples{
# stepwise backward regression
model <- lm(y ~ ., data = surgical)
ols_step_backward_aic(model)

# stepwise backward regression plot
model <- lm(y ~ ., data = surgical)
k <- ols_step_backward_aic(model)
plot(k)

# final model
k$model

# include or exclude variable
# force variables to be included in the selection process
ols_step_backward_aic(model, include = c("alc_mod", "gender"))

# use index of variable instead of name
ols_step_backward_aic(model, include = c(7, 6))

# force variable to be excluded from selection process
ols_step_backward_aic(model, exclude = c("alc_heavy", "bcs"))

# use index of variable instead of name
ols_step_backward_aic(model, exclude = c(8, 1))

}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.
}
\seealso{
Other variable selection procedures: 
\code{\link{ols_step_all_possible}()},
\code{\link{ols_step_backward_p}()},
\code{\link{ols_step_best_subset}()},
\code{\link{ols_step_both_aic}()},
\code{\link{ols_step_forward_aic}()},
\code{\link{ols_step_forward_p}()},
\code{\link{ols_step_rsquared}()}
}
\concept{variable selection procedures}
