% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-steparsq-both-regression.R
\name{ols_step_both_arsq}
\alias{ols_step_both_arsq}
\alias{plot.ols_step_both_arsq}
\alias{ols_steparsq_both}
\title{Stepwise Adjusted R-Squared regression}
\usage{
ols_step_both_arsq(model, progress = FALSE, details = FALSE)

\method{plot}{ols_step_both_arsq}(x, print_plot = TRUE, ...)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{progress}{Logical; if \code{TRUE}, will display variable selection progress.}

\item{details}{Logical; if \code{TRUE}, details of variable selection will
be printed on screen.}

\item{x}{An object of class \code{ols_step_both_arsq}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}

\item{...}{Other arguments.}
}
\value{
\code{ols_step_both_arsq} returns an object of class \code{"ols_step_both_arsq"}.
An object of class \code{"ols_step_both_arsq"} is a list containing the
following components:

\item{model}{model with the highest ARSQ; an object of class \code{lm}}
\item{predictors}{variables added/removed from the model}
\item{method}{addition/deletion}
\item{aics}{akaike information criteria}
\item{ess}{error sum of squares}
\item{rss}{regression sum of squares}
\item{rsq}{rsquare}
\item{arsq}{adjusted rsquare}
\item{steps}{total number of steps}
}
\description{
Build regression model from a set of candidate predictor variables by
entering and removing predictors based on adjusted R-squared, in a
stepwise manner until there is no variable left to enter or remove any more.
}
\section{Deprecated Function}{

\code{ols_steparsq_both()} has been deprecated. Instead use \code{ols_step_both_arsq()}.
}

\examples{
\dontrun{
# stepwise regression
model <- lm(y ~ ., data = stepdata)
ols_step_both_arsq(model)

# stepwise regression plot
model <- lm(y ~ ., data = stepdata)
k <- ols_step_both_arsq(model)
plot(k)

# final model
k$model

}
}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.
}
\seealso{
Other variable selection procedures: 
\code{\link{ols_step_all_possible}()},
\code{\link{ols_step_backward_aic}()},
\code{\link{ols_step_backward_arsq}()},
\code{\link{ols_step_backward_p}()},
\code{\link{ols_step_best_subset}()},
\code{\link{ols_step_both_aic}()},
\code{\link{ols_step_forward_aic}()},
\code{\link{ols_step_forward_arsq}()},
\code{\link{ols_step_forward_p}()}
}
\concept{variable selection procedures}
