% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-steparsq-forward-regression.R
\name{ols_step_forward_arsq}
\alias{ols_step_forward_arsq}
\alias{ols_step_forward_arsq.default}
\alias{plot.ols_step_forward_arsq}
\alias{ols_steparsq_forward}
\title{Stepwise Adjusted R-Squared forward regression}
\usage{
ols_step_forward_arsq(model, ...)

\method{ols_step_forward_arsq}{default}(
  model,
  max_steps = NULL,
  progress = FALSE,
  details = FALSE,
  ...
)

\method{plot}{ols_step_forward_arsq}(x, print_plot = TRUE, ...)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{...}{Other arguments.}

\item{max_steps}{Integer; if not \code{NULL}, will specify the maximum number of steps.}

\item{progress}{Logical; if \code{TRUE}, will display variable selection progress.}

\item{details}{Logical; if \code{TRUE}, will print the regression result at
each step.}

\item{x}{An object of class \code{ols_step_forward_arsq}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\value{
\code{ols_step_forward_arsq} returns an object of class \code{"ols_step_forward_arsq"}.
An object of class \code{"ols_step_forward_arsq"} is a list containing the
following components:

\item{model}{model with the highest ARSQ; an object of class \code{lm}}
\item{steps}{total number of steps}
\item{predictors}{variables added to the model}
\item{aics}{akaike information criteria}
\item{ess}{error sum of squares}
\item{rss}{regression sum of squares}
\item{rsq}{rsquare}
\item{arsq}{adjusted rsquare}
}
\description{
Build regression model from a set of candidate predictor variables by
entering predictors based on adjusted R-squared, in a stepwise
manner until there is no variable left to enter any more, or a maximum step has been reached.
}
\section{Deprecated Function}{

\code{ols_steparsq_forward()} has been deprecated. Instead use \code{ols_step_forward_arsq()}.
}

\examples{
# stepwise forward regression
model <- lm(y ~ ., data = surgical)
ols_step_forward_arsq(model)

# stepwise forward regression plot
model <- lm(y ~ ., data = surgical)
k <- ols_step_forward_arsq(model)
plot(k)

# final model
k$model


}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.
}
\seealso{
Other variable selection procedures: 
\code{\link{ols_step_all_possible}()},
\code{\link{ols_step_backward_aic}()},
\code{\link{ols_step_backward_arsq}()},
\code{\link{ols_step_backward_p}()},
\code{\link{ols_step_best_subset}()},
\code{\link{ols_step_both_aic}()},
\code{\link{ols_step_both_arsq}()},
\code{\link{ols_step_forward_aic}()},
\code{\link{ols_step_forward_p}()}
}
\concept{variable selection procedures}
