% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-step-rsquare.R
\name{ols_step_rsquared}
\alias{ols_step_rsquared}
\alias{ols_step_rsquared.default}
\alias{plot.ols_step_rsquared}
\title{Stepwise R-Square regression}
\usage{
ols_step_rsquared(model, ...)

\method{ols_step_rsquared}{default}(
  model,
  direction = c("forward", "backward", "both"),
  include = NULL,
  exclude = NULL,
  progress = FALSE,
  details = FALSE,
  ...
)

\method{plot}{ols_step_rsquared}(x, print_plot = TRUE, details = TRUE, ...)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{...}{Other arguments.}

\item{direction}{Stepwise direction; one among \code{forward}, \code{backward} or \code{both}.}

\item{include}{Character or numeric vector; variables to be included in selection process.}

\item{exclude}{Character or numeric vector; variables to be excluded from selection process.}

\item{progress}{Logical; if \code{TRUE}, will display variable selection progress.}

\item{details}{Logical; if \code{TRUE}, will print the regression result at
each step.}

\item{x}{An object of class \code{ols_step_rsquared}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\value{
\code{ols_step_rsquared} returns an object of class \code{"ols_step_rsquared"}.
An object of class \code{"ols_step_rsquared"} is a list containing the
following components:

\item{model}{final model; an object of class \code{lm}}
\item{metrics}{selection metrics}
\item{others}{list; info used for plotting and printing}
}
\description{
Build regression model from a set of candidate predictor variables by
entering predictors based on r-square and adjusted r-square, in a stepwise
manner until there is no variable left to enter any more.
}
\examples{
\dontrun{
# stepwise regression
model <- lm(y ~ ., data = stepdata)
ols_step_rsquared(model)

# stepwise regression plot
model <- lm(y ~ ., data = stepdata)
k <- ols_step_rsquared(model)
plot(k)

# final model
k$model

# include or exclude variables
# force variable to be included in selection process
model <- lm(y ~ ., data = stepdata)

ols_step_rsquared(model, include = c("x6"))

# use index of variable instead of name
ols_step_rsquared(model, include = c(6))

# force variable to be excluded from selection process
ols_step_rsquared(model, exclude = c("x2"))

# use index of variable instead of name
ols_step_rsquared(model, exclude = c(2))

# include & exclude variables in the selection process
ols_step_rsquared(model, include = c("x6"), exclude = c("x2"))

# use index of variable instead of name
ols_step_rsquared(model, include = c(6), exclude = c(2))
}

}
\seealso{
Other variable selection procedures: 
\code{\link{ols_step_all_possible}()},
\code{\link{ols_step_backward_aic}()},
\code{\link{ols_step_backward_p}()},
\code{\link{ols_step_best_subset}()},
\code{\link{ols_step_both_aic}()},
\code{\link{ols_step_forward_aic}()},
\code{\link{ols_step_forward_p}()}
}
\concept{variable selection procedures}
