% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-outlier-test.R
\name{ols_test_outlier}
\alias{ols_test_outlier}
\title{Bonferroni Outlier Test}
\usage{
ols_test_outlier(model, cut_off = 0.05, n_max = 10, ...)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{cut_off}{Bonferroni p-values cut off for reporting observations.}

\item{n_max}{Maximum number of observations to report, default is 10.}

\item{...}{Other arguments.}
}
\description{
Detect outliers using Bonferroni p values.
}
\examples{
# model
model <- lm(y ~ ., data = surgical)
ols_test_outlier(model)

}
