% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-version.R
\name{ols_update}
\alias{ols_update}
\title{Update olsrr package}
\usage{
ols_update(owner = "rsquaredacademy", repo = "olsrr")
}
\description{
Check whether olsrr is up-to-date, and will install
after confirmation.
}
\examples{
\dontrun{
ols_update()
}
}
