% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs-vs-pred-plot.R
\name{ovsp_plot}
\alias{ovsp_plot}
\title{Observed vs Predicted Plot}
\usage{
ovsp_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Plot of observed vs fitted values to assess the fit of the model.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ovsp_plot(model)
}

