% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-criteria.R
\name{pc}
\alias{pc}
\title{PC}
\usage{
pc(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
PC
}
\description{
Amemiya's prediction error
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
pc(model)
}

