% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred-diag.R
\name{pred_diag}
\alias{pred_diag}
\title{Visualize Response Variable}
\usage{
pred_diag(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Panel of plots to explore and visualize the response variable.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
pred_diag(model)
}

