% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence-measures.R
\name{press}
\alias{press}
\title{PRESS Statistic}
\usage{
press(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
PRESS Statistic
}
\description{
PRESS tells you how well the model will predict new data.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
press(model)
}

