% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure-error-anova.R
\name{pure_error_anova}
\alias{pure_error_anova}
\title{Lack  of Fit F Test}
\usage{
pure_error_anova(model, ...)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{...}{other parameters}
}
\value{
\code{pure_error_anova} returns an object of class
\code{"pure_error_anova"}. An object of class \code{"pure_error_anova"} is a
list containing the following components:

\item{lackoffit}{f statistic}
\item{pure_error}{pure error}
\item{rss}{regression sum of squares}
\item{ess}{error sum of squares}
\item{total}{total sum of squares}
\item{rms}{}
\item{ems}{p-value of \code{fstat}}
\item{lms}{degrees of freedom}
\item{pms}{name(s) of \code{variable}}
\item{rf}{name of \code{group_var}}
\item{lf}{f statistic}
\item{pr}{p-value of \code{fstat}}
\item{pl}{degrees of freedom}
\item{mpred}{name(s) of \code{variable}}
\item{df_rss}{name of \code{group_var}}
\item{df_ess}{f statistic}
\item{df_lof}{p-value of \code{fstat}}
\item{df_error}{degrees of freedom}
\item{final}{name(s) of \code{variable}}
\item{resp}{name of \code{group_var}}
\item{preds}{name of \code{group_var}}
}
\description{
Lack of Fit F Test
}
\details{
Some statistical tests, for example the analysis of variance, assume
that variances are equal across groups or samples. The Bartlett test can be
used to verify that assumption. Bartlett's test is sensitive to departures
from normality. That is, if your samples come from non-normal distributions,
then Bartlett's test may simply be testing for non-normality. The Levene test
is an alternative to the Bartlett test that is less sensitive to departures
from normality.
}

