% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual-qqplot.R
\name{qqresid}
\alias{qqresid}
\title{Residual QQ Plot}
\usage{
qqresid(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Graph for detecting violation of normality assumption.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
qqresid(model)
}

