% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted-line-properties.R
\name{reg_line}
\alias{reg_line}
\title{Regression Line}
\usage{
reg_line(response, predictor)
}
\arguments{
\item{response}{response variable}

\item{predictor}{explanatory variable}
}
\value{
\code{reg_line} returns  a list containing the
following components:

\item{mean_pred}{mean of \code{predictor}}
\item{mean_resp}{mean of \code{response}}
\item{model}{an object of class \code{lm}}
}
\description{
regression line always passes through xbar and ybar
}

