% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid-fit-spread-plot.R
\name{rfs_plot}
\alias{rfs_plot}
\title{Residual Fit Spread Plot}
\usage{
rfs_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{rfs_plot} returns a list containing the
following components:

\item{residuals}{f cook's d statistic}
\item{fitted_values}{f cook's d statistic}
\item{avg_fitted}{threshold for outliers}
\item{dev_fitted}{f cook's d statistic}
\item{ecdf_avg_fitted}{threshold for outliers}
\item{ecdf_resid}{f cook's d statistic}
}
\description{
Residual Fit Spread Plot
}

