% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivers.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{Rivers Data Set}
\format{A data frame with 20 rows and 6 variables:
\describe{
  \item{River}{id of the student}
  \item{Agr}{gender of the student}
  \item{Forest}{ethnic background of the student}
  \item{Rsdntial}{socio-economic status of the student}
  \item{ComIndl}{school type}
  \item{Nitrogen}{program type}
}}
\source{
\url{http://www.ats.ucla.edu/stat/spss/whatstat/whatstat.htm}
}
\usage{
rivers
}
\description{
A dataset containing information about rivers.
}
\keyword{datasets}

