% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid-fit-spread-plot.R
\name{rsd_plot}
\alias{rsd_plot}
\title{Residual Spread Plot}
\usage{
rsd_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{rsd_plot} returns a list containing the
following components:

\item{residuals}{f cook's d statistic}
\item{ecdf_resid}{threshold for outliers}
}
\description{
Residual Spread Plot
}

