% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid-vs-pred-plot.R
\name{rvsp_plot}
\alias{rvsp_plot}
\title{Residual vs Fitted Plot}
\usage{
rvsp_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Plot to detect non-linearity, unequal error variances, and outliers.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
rvsp_plot(model)
}

