% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-criteria.R
\name{sbc}
\alias{sbc}
\title{SBC}
\usage{
sbc(model, method = c("R", "STATA", "SAS"))
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{method}{a character vector; specify the method to compute AIC. Valid
options include R, STATA and SAS}
}
\value{
Bayesian Information Criteria
}
\description{
Bayesian Information Criteria
}
\examples{
# using R computation method
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
sbc(model)

# using STATA computation method
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
sbc(model, method = 'STATA')

# using SAS computation method
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
sbc(model, method = 'SAS')
}

