% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test.R
\name{score_test}
\alias{score_test}
\title{Score Test for heteroskedasticity}
\usage{
score_test(model, fitted.values = TRUE, rhs = FALSE, vars = NULL)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{fitted.values}{logical; if TRUE, use fitted values of regression model}

\item{rhs}{logical; if TRUE, specifies that tests for heteroskedasticity be
performed for the right-hand-side (explanatory) variables of the fitted
regression model}

\item{vars}{variables to be used for for heteroskedasticity test}
}
\value{
\code{score_test} returns an object of class \code{"score_test"}.
An object of class \code{"score_test"} is a list containing the
following components:

\item{score}{f statistic}
\item{p}{p value of \code{score}}
\item{df}{degrees of freedom}
\item{fv}{fitted values of the regression model}
\item{rhs}{name of explanatory variables of fitted regression model}
\item{resp}{response variable}
\item{preds}{predictors}
}
\description{
Test for heteroskedasticity
}

