% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-criteria.R
\name{sp}
\alias{sp}
\title{SP}
\usage{
sp(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
SP
}
\description{
Average prediction mean squared error
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
sp(model)
}

