% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stud-resid-plot.R
\name{srplot}
\alias{srplot}
\title{Studentized Residual Plot}
\usage{
srplot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{srplot} returns a list containing the
following components:

\item{dstudresid}{f cook's d statistic}
\item{threshold}{threshold for outliers}
\item{normal}{threshold for outliers}
\item{outlier}{f cook's d statistic}
}
\description{
Studentized Residual Plot
}

