% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepaic-backward.R
\name{stepaic_backward}
\alias{stepaic_backward}
\title{Stepwise AIC Backward Regression}
\usage{
stepaic_backward(model, ...)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{...}{other arguments}
}
\value{
\code{step_backward} returns an object of class \code{"step_backward"}.
An object of class \code{"step_backward"} is a list containing the
following components:

\item{steps}{f statistic}
\item{predictors}{p value of \code{score}}
\item{aics}{degrees of freedom}
\item{ess}{fitted values of the regression model}
\item{rss}{name of explanatory variables of fitted regression model}
\item{rsq}{response variable}
\item{arsq}{predictors}
}
\description{
Stepwise AIC Backward Regression
}

