% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepaic-both.R
\name{stepaic_both}
\alias{stepaic_both}
\title{Stepwise AIC Regression}
\usage{
stepaic_both(model, details = FALSE)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{details}{logical; if TRUE details of variable selection will be printed on screen}
}
\value{
\code{stepaic_both} returns an object of class \code{"stepaic_both"}.
An object of class \code{"stepaic_both"} is a list containing the
following components:

\item{predictors}{variables retained in the model}
\item{method}{addition/deletion}
\item{aics}{akaike information criteria}
\item{ess}{error sum of squares}
\item{rss}{regression sum of squares}
\item{rsq}{rsquare}
\item{arsq}{adjusted rsquare}
\item{steps}{total number of steps}
}
\description{
Stepwise AIC Regression
}

