% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepaic-forward.R
\name{stepaic_forward}
\alias{stepaic_forward}
\title{Stepwise AIC Forward Regression}
\usage{
stepaic_forward(model, ...)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{...}{other arguments}
}
\value{
\code{stepaic_forward} returns an object of class \code{"stepaic_forward"}.
An object of class \code{"stepaic_forward"} is a list containing the
following components:

\item{steps}{f statistic}
\item{predictors}{p value of \code{score}}
\item{aics}{degrees of freedom}
\item{ess}{fitted values of the regression model}
\item{rss}{name of explanatory variables of fitted regression model}
\item{rsq}{response variable}
\item{arsq}{predictors}
}
\description{
Stepwise AIC Forward Regression
}

