% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-data-step.R
\docType{data}
\name{stepdata}
\alias{stepdata}
\title{Test Data Set}
\format{An object of class \code{data.frame} with 20000 rows and 7 columns.}
\usage{
stepdata
}
\description{
Test Data Set
}
\keyword{datasets}
