% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studres-chart.R
\name{studres_chart}
\alias{studres_chart}
\title{Studentized Residual Chart}
\usage{
studres_chart(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Chart for identifying outliers
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
studres_chart(model)
}

