% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/added-variable-plot.R
\name{addvar_plot}
\alias{addvar_plot}
\title{Added Variable Plot}
\usage{
addvar_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Show the effect of adding another variable to a model already
having one or more independent variables.
}
\details{
The added variable plot was introduced by Mosteller and Tukey (1977). It enables
us to visualize the regression coefficient of a new variable being considered to
be included in a model. The plot can be constructed for each predictor variable.

Let us assume we want to test the effect of adding/removing variable *X* from a
model. Let the response variable of the model be *Y*

Steps to construct an added variable plot:

- Regress *Y* on all variables other than *X* and store the residuals (*Y* residuals).
- Regress *X* on all the other variables inlcuded in the model (*X* residuals).
- Construct a scatter plot of *Y* residuals and *X* residuals.

What do the *Y* and *X* residuals represent? The *Y* residuals represent the part
of **Y** not explained by all the variables other than X. The *X* residuals
represent the part of **X** not explained by other variables. The slope of the line
fitted to the points in the added variable plot is equal to the regression
coefficient when **Y** is regressed on all variables including **X**.

A strong linear relationship in the added variable plot indicates the increased
importance of the contribution of **X** to the model already containing the
other predictors.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
addvar_plot(model)
}
\references{
Regression Analysis by Example, 5th Edition, Samprit Chatterjee
and Ali S Hadi, 2012, John Wiley & Sons
}

