% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/added-variable-plot.R
\name{addvar_plot}
\alias{addvar_plot}
\title{Added Variable Plot}
\usage{
addvar_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Show the effect of adding another variable to a model already
having one or more independent variables.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
addvar_plot(model)
}

