% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-criteria.R
\name{aic}
\alias{aic}
\title{AIC}
\usage{
aic(model, method = c("R", "STATA", "SAS"))
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{method}{a character vector; specify the method to compute AIC. Valid
options include R, STATA and SAS}
}
\value{
Akaike Information Criteria
}
\description{
Akaike Information Criteria
}
\examples{
# using R computation method
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
aic(model)

# using STATA computation method
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
aic(model, method = 'STATA')

# using SAS computation method
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
aic(model, method = 'SAS')
}

