% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartlett-test.R
\name{bartlett_test}
\alias{bartlett_test}
\title{Bartlett Test}
\usage{
bartlett_test(variable, ...)
}
\arguments{
\item{variable}{a numeric vector}

\item{...}{numeric vectors}
}
\value{
\code{bartlett_test} returns an object of class \code{"bartlett_test"}.
An object of class \code{"bartlett_test"} is a list containing the
following components:

\item{fstat}{f statistic}
\item{pval}{p-value of \code{fstat}}
\item{df}{degrees of freedom}
\item{var_c}{name(s) of \code{variable}}
\item{g_var}{name of \code{group_var}}
}
\description{
Test if k samples have equal variances
}
\details{
Some statistical tests, for example the analysis of variance, assume
that variances are equal across groups or samples. The Bartlett test can be
used to verify that assumption. Bartlett's test is sensitive to departures
from normality. That is, if your samples come from non-normal distributions,
then Bartlett's test may simply be testing for non-normality. The Levene test
is an alternative to the Bartlett test that is less sensitive to departures
from normality.
}

