% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-data-cement.R
\docType{data}
\name{cement}
\alias{cement}
\title{Test Data Set}
\format{
An object of class \code{data.frame} with 13 rows and 6 columns.
}
\usage{
cement
}
\description{
Test Data Set
}
\keyword{internal}
