% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coll-diag.R
\name{coll_diag}
\alias{coll_diag}
\alias{eigen_cindex}
\alias{vif_tol}
\title{Collinearity Diagnostics}
\usage{
coll_diag(model)

vif_tol(model)

eigen_cindex(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{coll_diag} returns an object of class \code{"coll_diag"}.
An object of class \code{"coll_diag"} is a list containing the
following components:

\item{vif_t}{tolerance and variance inflation factors}
\item{eig_cindex}{eigen values and condition index}
}
\description{
Collinearity implies two variables are near perfect linear combinations of one
another. Multicollinearity involves more than two variables. In the presence of multicollinearity, 
regression estimates are unstable and have high standard errors. \code{coll_diag}  
returns variance inflation factor, tolerance and condition indices. Collinearity is spotted by 
finding 2 or more variables that have large proportions of variance (.50 or more) that correspond 
to large condition indices. A rule of thumb is to label as large those condition indices in the range of 30 or larger.
Big values of VIF and small values of Tolerance indicate multicollinearity.
}
\references{
Belsley, D. A., Kuh, E., and Welsch, R. E. (1980). Regression Diagnostics: Identifying Influential Data and
Sources of Collinearity. New York: John Wiley & Sons.
}

