% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks-d-barplot.R
\name{cooksd_bplot}
\alias{cooksd_bplot}
\title{Cooks' D Bar Plot}
\usage{
cooksd_bplot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Bar Plot of Cook's distance to detect observations that strongly influence fitted values of the model.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
cooksd_bplot(model)
}

