% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks-d-chart.R
\name{cooksd_chart}
\alias{cooksd_chart}
\title{Cooks' d Chart}
\usage{
cooksd_chart(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{cooksd_chart} returns a list containing the
following components:

\item{cooks_d}{f cook's d statistic}
\item{threshold}{threshold for outliers}
\item{outliers}{residual points that are outliers}
}
\description{
Cooks' d Chart
}

