% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm-test.R
\name{corr_test}
\alias{corr_test}
\title{Correlation Test}
\usage{
corr_test(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
correlation between fitted regression model residuals and expected
values of residuals
}
\description{
Correlation Test
}
\details{
Some statistical tests, for example the analysis of variance, assume
that variances are equal across groups or samples. The Bartlett test can be
used to verify that assumption. Bartlett's test is sensitive to departures
from normality. That is, if your samples come from non-normal distributions,
then Bartlett's test may simply be testing for non-normality. The Levene test
is an alternative to the Bartlett test that is less sensitive to departures
from normality.
}

