% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-plus-resid-plot.R
\name{cplusr_plot}
\alias{cplusr_plot}
\title{Component Plus Residual Plot}
\usage{
cplusr_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Component plus residual plot
}

