% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffits-plot.R
\name{dffits_plot}
\alias{dffits_plot}
\title{DFFITS Plot}
\usage{
dffits_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{dffits_plot} returns a list containing the
following components:

\item{dffits}{dffits}
\item{threshold}{threshold for outliers}
\item{outliers}{residual points that are outliers}
}
\description{
DFFITS Plot
}

