% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsresid-vs-pred-plot.R
\name{dsrvsp_plot}
\alias{dsrvsp_plot}
\title{Deleted Studentized Residual vs Predictor Plot}
\usage{
dsrvsp_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{dsrvsp_plot} returns a list containing the
following components:

\item{fitted}{fitted values of the regression model}
\item{dstudresid}{deleted studentized residuals}
\item{threshold}{threshold for outliers}
\item{normal}{normal residual points}
\item{outliers}{residual points that are outliers}
}
\description{
Deleted Studentized Residual vs Predictor Plot
}

