% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-data-fitness.R
\docType{data}
\name{fitness}
\alias{fitness}
\title{Test Data Set}
\format{An object of class \code{data.frame} with 31 rows and 7 columns.}
\usage{
fitness
}
\description{
Test Data Set
}
\keyword{datasets}

