% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual-hist.R
\name{hist_resid}
\alias{hist_resid}
\title{Residual Histogram}
\usage{
hist_resid(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{hist_resid} returns a list containing the
following components:

\item{residuals}{f cook's d statistic}
\item{breaks}{f cook's d statistic}
\item{counts}{threshold for outliers}
\item{density}{f cook's d statistic}
\item{mids}{threshold for outliers}
\item{xname}{f cook's d statistic}
\item{equidist}{f cook's d statistic}
\item{norm}{f cook's d statistic}
}
\description{
Residual Histogram
}

