% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm-test.R
\name{norm_test}
\alias{norm_test}
\title{Test for normality}
\usage{
norm_test(y, ...)
}
\arguments{
\item{y}{a numeric vector}

\item{...}{other arguments}
}
\value{
\code{norm_test} returns an object of class \code{"norm_test"}.
An object of class \code{"norm_test"} is a list containing the
following components:

\item{kolmogorv}{kolmogorv smirnov statistic}
\item{shapiro}{shapiro wilk statistic}
\item{cramer}{cramer von mises statistic}
\item{anderson}{anderson darling statistic}
}
\description{
Test for detecting violation of normality assumption.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
norm_test(model)
}

