% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-bartlett-test.R
\name{ols_bartlett_test}
\alias{ols_bartlett_test}
\alias{ols_bartlett_test.default}
\title{Bartlett Test}
\usage{
ols_bartlett_test(data, ...)

\method{ols_bartlett_test}{default}(data, ..., group_var = NULL)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble}}

\item{...}{columns in \code{data}}

\item{group_var}{grouping variable}
}
\value{
\code{ols_bartlett_test} returns an object of class \code{"ols_bartlett_test"}.
An object of class \code{"ols_bartlett_test"} is a list containing the
following components:

\item{fstat}{f statistic}
\item{pval}{p-value of \code{fstat}}
\item{df}{degrees of freedom}
}
\description{
Test if k samples are from populations with equal variances.
}
\details{
Bartlett's test is used to test if variances across samples is equal.
It is sensitive to departures from normality. The Levene test
is an alternative test that is less sensitive to departures from normality.
}
\examples{
\dontrun{
# using grouping variable
library(descriptr)
ols_bartlett_test(mtcarz, mpg, group_var = cyl)
}

\dontrun{
# using variables
ols_bartlett_test(hsb, read, write)
}

}
\references{
Snedecor, George W. and Cochran, William G. (1989), Statistical Methods,
Eighth Edition, Iowa State University Press.
}
