% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-collinearity-diagnostics.R
\name{ols_coll_diag}
\alias{ols_coll_diag}
\alias{ols_eigen_cindex}
\alias{ols_vif_tol}
\title{Collinearity Diagnostics}
\usage{
ols_coll_diag(model)

ols_vif_tol(model)

ols_eigen_cindex(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{ols_coll_diag} returns an object of class \code{"ols_coll_diag"}.
An object of class \code{"ols_coll_diag"} is a list containing the
following components:

\item{vif_t}{tolerance and variance inflation factors}
\item{eig_cindex}{eigen values and condition index}
}
\description{
Variance inflation factor, tolerance, eigenvalues and condition indices.
}
\examples{
# model
model <- lm(mpg ~ disp + hp + wt + drat, data = mtcars)

# vif and tolerance
ols_vif_tol(model)

# eigenvalues and condition indices
ols_eigen_cindex(model)

# collinearity diagnostics
ols_coll_diag(model)

}
\references{
Belsley, D. A., Kuh, E., and Welsch, R. E. (1980). Regression Diagnostics: Identifying Influential Data and
Sources of Collinearity. New York: John Wiley & Sons.
}

