% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-cooks-d-barplot.R
\name{ols_cooksd_barplot}
\alias{ols_cooksd_barplot}
\title{Cooks' D Bar Plot}
\usage{
ols_cooksd_barplot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Bar Plot of Cook's distance to detect observations that strongly influence fitted values of the model.
}
\details{
Cook's distance was introduced by American statistician R Dennis Cook in 1977. It is used 
to identify influential data points. It depends on both the residual and leverage i.e it takes it account
both the \emph{x} value and \emph{y} value of the observation.

Steps to compute Cook's distance:

\itemize{
  \item Delete observations one at a time.
  \item Refit the regression model on remaining \eqn{n - 1} observations
  \item examine how much all of the fitted values change when the ith observation is deleted.
}

A data point having a large cook's d indicates that the data point strongly influences the fitted values.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_cooksd_barplot(model)
}

