% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-dsresid-vs-pred-plot.R
\name{ols_dsrvsp_plot}
\alias{ols_dsrvsp_plot}
\title{Deleted Studentized Residual vs Fitted Values Plot}
\usage{
ols_dsrvsp_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Plot for detecting outliers.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_dsrvsp_plot(model)
}

