% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-influence-measures.R
\name{ols_hadi}
\alias{ols_hadi}
\title{Hadi's Measure}
\usage{
ols_hadi(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
hadi's measure
}
\description{
Measure of influence based on the fact that influential observations in either
the response variable or in the predictors or both.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_hadi(model)
}
\references{
Chatterjee, Samprit and Hadi, Ali. Regression Analysis by Example. 5th ed. N.p.: John Wiley & Sons, 2012. Print.
}
