% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-diagnostics-panel.R
\name{ols_plot_diagnostics}
\alias{ols_plot_diagnostics}
\title{Diagnostics panel}
\usage{
ols_plot_diagnostics(model, print_plot = TRUE)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\description{
Panel of plots for regression diagnostics.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_plot_diagnostics(model)

}
