% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-diagnostics-panel.R
\name{ols_plot_diagnostics}
\alias{ols_plot_diagnostics}
\alias{ols_diagnostic_panel}
\title{Diagnostics panel}
\usage{
ols_plot_diagnostics(model)
}
\arguments{
\item{model}{An object of class \code{lm}.

#' @section Deprecated Function:
\code{ols_diagnostic_panel()} has been deprecated. Instead use \code{ols_plot_diagnostics()}.}
}
\description{
Panel of plots for regression diagnostics.
}
\examples{
\dontrun{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_plot_diagnostics(model)
}

}
