% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-residual-boxplot.R
\name{ols_plot_resid_box}
\alias{ols_plot_resid_box}
\alias{ols_rsd_boxplot}
\title{Residual box plot}
\usage{
ols_plot_resid_box(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Box plot of residuals to examine if residuals are normally distributed.
}
\section{Deprecated Function}{

\code{ols_rsd_boxplot()} has been deprecated. Instead use \code{ols_plot_resid_box()}.
}

\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_plot_resid_box(model)

}
\seealso{
Other residual diagnostics: \code{\link{ols_plot_resid_fit}},
  \code{\link{ols_plot_resid_hist}},
  \code{\link{ols_plot_resid_qq}},
  \code{\link{ols_test_correlation}},
  \code{\link{ols_test_normality}}
}
