% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-dsresid-vs-pred-plot.R,
%   R/ols-residual-vs-predicted-plot.R
\name{ols_plot_resid_fit}
\alias{ols_plot_resid_fit}
\alias{ols_dsrvsp_plot}
\alias{ols_plot_resid_fit}
\alias{ols_rvsp_plot}
\title{Deleted studentized residual vs fitted values plot}
\usage{
ols_plot_resid_fit(model)

ols_plot_resid_fit(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{model}{An object of class \code{lm}.}
}
\value{
\code{ols_plot_resid_fit} returns  a list containing the
following components:

\item{outliers}{a tibble with observation number, fitted values and deleted studentized
residuals that exceed the \code{threshold} for classifying observations as
outliers/influential observations}
\item{threshold}{\code{threshold} for classifying an observation as an outlier/influential observation}
}
\description{
Plot for detecting violation of assumptions about residuals such as
non-linearity, constant variances and outliers. It can also be used to
examine model fit.

Scatter plot of residuals on the y axis and fitted values on the
x axis to detect non-linearity, unequal error variances, and outliers.
}
\details{
Studentized deleted residuals (or externally studentized residuals) is the
deleted residual divided by its estimated standard deviation. Studentized
residuals are going to be more effective for detecting outlying Y
observations than standardized residuals. If an observation has an externally
studentized residual that is larger than 2 (in absolute value) we can call
it an outlier.

Characteristics of a well behaved residual vs fitted plot:

\itemize{
  \item The residuals spread randomly around the 0 line indicating that the relationship is linear.
  \item The residuals form an approximate horizontal band around the 0 line indicating homogeneity of error variance.
  \item No one residual is visibly away from the random pattern of the residuals indicating that there are no outliers.
}
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_plot_resid_fit(model)

model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_plot_resid_fit(model)

}
\seealso{
[ols_rsdlev_plot()], [ols_srsd_plot()], [ols_srsd_chart()]

Other residual diagnostics: \code{\link{ols_plot_resid_box}},
  \code{\link{ols_plot_resid_hist}},
  \code{\link{ols_plot_resid_qq}},
  \code{\link{ols_test_correlation}},
  \code{\link{ols_test_normality}}
}
