% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-rstud-vs-lev-plot.R
\name{ols_plot_resid_lev}
\alias{ols_plot_resid_lev}
\title{Studentized residuals vs leverage plot}
\usage{
ols_plot_resid_lev(model, threshold = NULL, print_plot = TRUE)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{threshold}{Threshold for detecting outliers. Default is 2.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\description{
Graph for detecting outliers and/or observations with high leverage.
}
\examples{
model <- lm(read ~ write + math + science, data = hsb)
ols_plot_resid_lev(model)
ols_plot_resid_lev(model, threshold = 3)

}
\seealso{
\code{\link[=ols_plot_resid_stud_fit]{ols_plot_resid_stud_fit()}}, \code{\link[=ols_plot_resid_lev]{ols_plot_resid_lev()}}
}
