% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-potential-residual-plot.R
\name{ols_potrsd_plot}
\alias{ols_potrsd_plot}
\title{Potential Residual Plot}
\usage{
ols_potrsd_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Plot to aid in classifying unusual observations as high-leverage points,
outliers, or a combination of both.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_potrsd_plot(model)
}
\references{
Chatterjee, Samprit and Hadi, Ali. Regression Analysis by Example. 5th ed. N.p.: John Wiley & Sons, 2012. Print.
}

