% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-influence-measures.R
\name{ols_pred_rsq}
\alias{ols_pred_rsq}
\title{Predicted Rsquare}
\usage{
ols_pred_rsq(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
Predicted Rsquare
}
\description{
Use predicted rsquared to determine how well the model predicts responses for new observations. 
Larger values of predicted R2 indicate models of greater predictive ability.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_pred_rsq(model)
}
