% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-plots-data.R
\name{ols_prep_cdplot_data}
\alias{ols_prep_cdplot_data}
\title{Cooks' D plot data}
\usage{
ols_prep_cdplot_data(model)
}
\description{
Prepare data for cook's d bar plot.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_prep_cdplot_data(model)

}
