% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-influence-measures.R
\name{ols_press}
\alias{ols_press}
\title{PRESS Statistic}
\usage{
ols_press(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
PRESS Statistic
}
\description{
PRESS tells you how well the model will predict new data.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_press(model)
}
\references{
Kutner, MH, Nachtscheim CJ, Neter J and Li W., 2004, Applied Linear Statistical Models (5th edition). 
Chicago, IL., McGraw Hill/Irwin.
}

