% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-fitted-line-properties.R
\name{ols_reg_line}
\alias{ols_reg_line}
\title{Simple linear regression line}
\usage{
ols_reg_line(response, predictor)
}
\arguments{
\item{response}{Response variable.}

\item{predictor}{Predictor variable.}
}
\description{
Plot to demonstrate that the regression line always passes  through mean of
the response and predictor variables.
}
\examples{
ols_reg_line(mtcars$mpg, mtcars$disp)

}
