% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-residual-boxplot.R
\name{ols_rsd_boxplot}
\alias{ols_rsd_boxplot}
\title{Residual Box Plot}
\usage{
ols_rsd_boxplot(model)
}
\arguments{
\item{model}{an object of class \code{lm}
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_rsd_boxplot(model)}
}
\description{
Box plot of residuals
}

