% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-residual-qqplot.R
\name{ols_rsd_qqplot}
\alias{ols_rsd_qqplot}
\title{Residual QQ Plot}
\usage{
ols_rsd_qqplot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Graph for detecting violation of normality assumption.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
ols_rsd_qqplot(model)
}

