% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-information-criteria.R
\name{ols_sbic}
\alias{ols_sbic}
\title{SBIC}
\usage{
ols_sbic(model, full_model)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{full_model}{an object of class \code{lm}
options include R, STATA and SAS}
}
\value{
Sawa's Bayesian Information Criteria
}
\description{
Sawa's Bayesian Information Criteria
}
\examples{
full_model <- lm(mpg ~ ., data = mtcars)
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_sbic(model, full_model)
}

