% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-normality-tests.R
\name{ols_test_correlation}
\alias{ols_test_correlation}
\alias{ols_corr_test}
\title{Correlation test for normality}
\usage{
ols_test_correlation(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
Correlation between fitted regression model residuals and expected
values of residuals
}
\description{
Correlation between observed residuals and expected residuals under normality.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_test_correlation(model)

}
\seealso{
Other residual diagnostics: \code{\link{ols_plot_resid_box}},
  \code{\link{ols_rsd_hist}}, \code{\link{ols_rsd_qqplot}},
  \code{\link{ols_rvsp_plot}},
  \code{\link{ols_test_normality}}
}
