% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olsrr.R
\docType{package}
\name{olsrr}
\alias{olsrr}
\title{\code{olsrr} package}
\description{
Tools for teaching and learning OLS regression
}
\details{
See the README on
\href{https://github.com/rsquaredacademy/olsrr}{GitHub}
}
\keyword{internal}
