% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential-resid-plot.R
\name{poten_resid_plot}
\alias{poten_resid_plot}
\title{Potential Residual Plot}
\usage{
poten_resid_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Plot to aid in classifying unusual observations as high-leverage points,
outliers, or a combination of both.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
poten_resid_plot(model)
}

