% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-fitted-line-properties.R
\name{reg_line}
\alias{reg_line}
\title{Regression Line}
\usage{
reg_line(response, predictor)
}
\arguments{
\item{response}{response variable}

\item{predictor}{explanatory variable}
}
\description{
Regression line always passes through xbar and ybar
}

