% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid-fit-spread-plot.R
\name{rfs_plot}
\alias{rfs_plot}
\title{Residual Fit Spread Plot}
\usage{
rfs_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Plot to detect non-linearity, influential observations and outliers.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
rfs_plot(model)
}

