% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-data-rivers.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{Test Data Set}
\format{An object of class \code{data.frame} with 20 rows and 6 columns.}
\usage{
rivers
}
\description{
Test Data Set
}
\keyword{datasets}

