% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid-vs-regressor-plot.R
\name{rvsr_plot}
\alias{rvsr_plot}
\title{Residual vs Regressors Plot}
\usage{
rvsr_plot(model, panel = TRUE)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{panel}{logical; if TRUE, creates a panel of plots}
}
\description{
Residual vs Regressors Plot
}

