% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-criteria.R
\name{sbic}
\alias{sbic}
\title{SBIC}
\usage{
sbic(model, full_model)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{full_model}{an object of class \code{lm}
options include R, STATA and SAS}
}
\value{
Sawa's Bayesian Information Criteria
}
\description{
Sawa's Bayesian Information Criteria
}
\examples{
full_model <- lm(mpg ~ ., data = mtcars)
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
sbic(model, full_model)
}

