% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-criteria.R
\name{sp}
\alias{sp}
\title{SP}
\usage{
sp(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
SP
}
\description{
Average prediction mean squared error
}
\details{
Some statistical tests, for example the analysis of variance, assume
that variances are equal across groups or samples. The Bartlett test can be
used to verify that assumption. Bartlett's test is sensitive to departures
from normality. That is, if your samples come from non-normal distributions,
then Bartlett's test may simply be testing for non-normality. The Levene test
is an alternative to the Bartlett test that is less sensitive to departures
from normality.
}

