% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stud-resid-plot.R
\name{srplot}
\alias{srplot}
\title{Studentized Residual Plot}
\usage{
srplot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Graph for identifying outliers
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
srplot(model)
}

