% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-forward.R
\name{step_forward}
\alias{step_forward}
\title{Stepwise Forward Regression}
\usage{
step_forward(model, ...)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{...}{other arguments}
}
\value{
\code{step_forward} returns an object of class \code{"step_forward"}.
An object of class \code{"step_forward"} is a list containing the
following components:

\item{steps}{f statistic}
\item{predictors}{p value of \code{score}}
\item{rsquare}{degrees of freedom}
\item{aic}{fitted values of the regression model}
\item{sbc}{name of explanatory variables of fitted regression model}
\item{sbic}{response variable}
\item{adjr}{predictors}
\item{rmse}{predictors}
\item{mallows_cp}{predictors}
\item{indvar}{predictors}
}
\description{
Build regression model from a set of candidate predictor variables by entering predictors based on 
p values, in a stepwise manner until there is no variable left to enter any more.
}
\examples{
# stepwise forward regression
model <- lm(y ~ ., data = surgical)
step_forward(model)

# stepwise forward regression plot
model <- lm(y ~ ., data = surgical)
k <- step_forward(model)
plot(k)

}

