% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.R
\name{stepwise}
\alias{stepwise}
\title{Stepwise Regression}
\usage{
stepwise(model, ...)
}
\arguments{
\item{model}{an object of class \code{lm}}

\item{...}{other arguments}
}
\value{
\code{stepwise} returns an object of class \code{"stepwise"}.
An object of class \code{"stepwise"} is a list containing the
following components:

\item{orders}{f statistic}
\item{method}{f statistic}
\item{steps}{f statistic}
\item{predictors}{p value of \code{score}}
\item{rsquare}{degrees of freedom}
\item{aic}{fitted values of the regression model}
\item{sbc}{name of explanatory variables of fitted regression model}
\item{sbic}{response variable}
\item{adjr}{predictors}
\item{rmse}{predictors}
\item{mallows_cp}{predictors}
\item{indvar}{predictors}
}
\description{
Stepwise Regression
}

