% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studres-chart.R
\name{studres_chart}
\alias{studres_chart}
\title{Studentized Residual Chart}
\usage{
studres_chart(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\value{
\code{studres_chart} returns a list containing the
following components:

\item{studresid}{f cook's d statistic}
\item{threshold}{threshold for outliers}
\item{outliers}{f cook's d statistic}
}
\description{
Studentized Residual Chart
}

