% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstud-vs-lev-plot.R
\name{studvslev_plot}
\alias{studvslev_plot}
\title{Studentized Residuals vs Leverage Plot}
\usage{
studvslev_plot(model)
}
\arguments{
\item{model}{an object of class \code{lm}}
}
\description{
Graph for detecting influential observations
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
studvslev_plot(model)
}

